convert image builder to a CLI app
Turn `create_windows_server_2022_image.sh` into a CLI app. This doesn't change
very much about how the imaging process functions (for now the app relies mostly
on shelling out to the same utilities the old script used to create and
partition disks and run VMs), but it does have a couple of small advantages:

- The app is built around a generic "script runner" that exectues an arbitrary
  set of steps and pretty-prints their names and outcomes. (At one point I had
  designs on creating a full-fledged TUI app for this, but stopped at a CLI app
  for now.)
- The app gets to use clap to parse command lines and generate help text.
- The app supports both illumos/Propolis and Linux/QEMU-based workflows and can
  share code between them. (The workflows are still somewhat mechanically
  different because Propolis doesn't yet support removable disk drives.)

In the future we can extend the app to, e.g., dynamically update unattend files
based on the Windows version the user is trying to image.

Other changes in this commit:

- Remove buildomat jobs to create new images. Buildomat outputs are publicly
  available, but we'd prefer not to distribute any Windows installation media
  ourselves. We create images infrequently enough that we can run the tools
  manually when we need new images; if this turns cumbersome we can find other
  ways to create images automatically.
- Slightly update install_prerequisites for illumos to pass the
  `-mgeneral_regs_only` flag when building the FUSE driver in order to work
  around a panic I was hitting caused by the driver using floating point
  registers unexpectedly.
- Clean up shellcheck warnings in install_prerequisites.
- Update the README to describe how to use the CLI app and the contents of the
  images it produces.

Tested by running the app on both illumos and Linux and verifying that it
produces a bootable Windows image. Examined the Windows image and verified
that it honored my cloud-init settings and that the relevant virtio drivers were
installed and loaded.