illumos: move OxidePrepBaseImage invocation & refactor
Refactor the illumos build_installation_disk script to use fewer closures and
look more like its counterparts.

Move the invocation of OxidePrepBaseImage.ps1 into a Windows batch script and
invoke that from Autounattend.xml's RunSynchronousCommand. This allows the XML
file not to contain any ampersands, which XML parsers treat as control
characters (but which Windows Setup interprets literally). This in turn will
allow the app to use the xml-rs crate to parse Autounattend.xml and update the
image it installs and/or the virtio drivers to inject.