fix illumos build; add interactive mode checks
Plug the illumos build into the UI added for Linux.

Add a notion of whether the program is being run in "interactive" mode (defaults
to whether stdout is a tty, but can be forced on or off). Gate some things on
the program's interactivity:

- Only display the `indicatif` UI in interactive mode
- Only ask the user to confirm settings in interactive mode
- Don't inherit stdout/stderr for long-running commands (like `7z`) in
  interactive mode (where the UI displays instead)

Currently non-interactive mode is something of a hack: the runner creates
progress bars irrespective of the program's mode and just suppresses their
display if the program is running non-interactively. A better way to deal with
this is to convert `struct Ui` into a trait, but this is left for a future
commit.