runner: warn instead of dying on missing unattend files
Allow users to proceed with image construction even if some of the unattend
files in `UNATTEND_FILES` are missing (but display warnings for these files in
the preflight checks). This allows the tool to create images that don't use all
of the default unattend files (e.g. if the user has customized the setup
scripts so that they don't install cloudbase-init, then the cloudbase-init
configuration files don't need to be copied).