WIP: consolidate OS-specific scripts
For the time being `prep.cmd` and `Autounattend.xml` are still environment-
specific, but the other scripts are either identical (cloudbase-init config
files, the specialize answer file) or are called by someone else who can pass
a parameter (OxidePrepBaseImage.ps1).

`install_prerequisites.sh` now works on either illumos or Linux (mostly by
moving the old OS-specific scripts into functions and deciding which one to
call from `uname` output).

This breaks the "point wimsy at a directory in the repo containing all the
unattend files" workflow, so the next step is to write a script that generates
an unattend directory from the checked-in components.