update README and helper scripts
Rewrite the README in anticipation of having release tarballs that contain all
the necessary programs and installation scripts in a single archive.

Also change the last section of the README so that it only describes how the
tool configures images in its default configuration without describing how an
image "should" be configured to work on an Oxide rack. The latter topic is
better suited for a guide in the public documentation.