Clean up OS-specific scripts (#5)
Instead of having two copies of all the unattend scripts, have one copy of
everything except Autounattend.xml and the prep.cmd batch script that invokes
OxidePrepBaseImage.ps1. This makes most of the interesting configuration in the
unattend files much easier to maintain.

Rework the README to reflect these changes. The new README refers to release
tarballs that don't exist yet--I'll work on that next but wanted to get the repo
reorganized before starting to work out how to package things.

Also add qemu-system-gui to the Linux prereqs list, since this now needs to be
installed explicitly to get the VGA console to work.