Write child command stdout/stderr to log files (#6)
It's sometimes useful to look at the stdout/stderr of the various processes
wimsy invokes. This is especially true for QEMU: wimsy configures it to write
data from the guest's COM1 to stdout, and the setup scripts direct
OxidePrepBaseImage.ps1's output to be written there.

Clean up the `Ui` trait a bit and change `run_command_check_status` to ensure
these outputs get logged for most command invocations.

Also fix a test bug that crept in with the last batch of changes (the locations
of the in-box Autounattend.xml files moved, but I forgot to update the tests to
point to their new homes).