pass --shrink to `qemu-img resize` (#22)
Modern versions of qemu-img require users to pass --shrink when shrinking a raw
disk with the resize subcommand, so try to pass that flag when shrinking a disk.
If this fails, fall back to invoking qemu-img resize without --shrink to
accommodate older versions of qemu-img that don't support the flag. (These
versions of qemu-img are probably not especially common, but they do exist: the
Debian 9 online man pages for qemu-img don't mention --resize, and the illumos
version of the tool in the system/kvm package doesn't implement it.)

Tested by running wimsy on an Ubuntu 20.04 workstation with qemu-img 4.2.1
(supports --shrink but doesn't require it).